47-commit

+++ src/services/categories.js
@@ -1,5 +1,6 @@
 const { httpError } = require("../util/http-error");
 const { Category } = require("../entities/category");
+const { Tutorial } = require("../entities/tutorial");
 
 const getCategories = async () => {
   return Category.find();
@@ -29,6 +30,10 @@ const updateCategory = async (id, name) => {
 };
 
 const deleteCategory = async (id) => {
+  const totalCount = await Tutorial.countDocuments({ category: id });
+  if (totalCount > 0) {
+    throw httpError(409, "Category has Tutorials");
+  }
   const category = await Category.findByIdAndDelete(id);
   if (!category) {
     throw httpError(404, "Category is not found");
